drug_discovery_classification_cyp450/requirements.txt
streamlit
rdkit-pypi
tensorflow
pandas
matplotlib
Pillow
pydot
